code: remove all the compiling warnings
logs: add further information when the transaction is completed or failed
logs: expand the vhost macro for redirects in logs
logs: add the farm name and thread id to every log
logs: add a log error when zproxy receives some signal
logs: log error if the SSL handshake fails
sessions: fix some errors regarding static sessions
sessions: return an error if the session ID is an empty string
sessions: maintain the sessions after reloading the zproxy process
stats: maintain the stats after a process reload
ctl: return the HTTP code 400 on failure
parser: add an error to the parser if the request JSON contains a double value with more than one dot. Example ("ip": 127.0.0.1)
algorithms: fix an error with the service priority. It did not increase when some backend changed the status
http: fix an issue matching services with header patterns that contain the "$" character
http: fix an issue downloading long files due to an overload saving the Content-Length header
http: clean extra HTTP headers for the next request when using HTTP version 1.1
http: respond HTTP error code if the response parsing failed
timeout: disrupt the connection if the backend did not send data for "timeout" seconds
sockets: return an error if after adding header the maximum header size is reached

+++ src/ssl/ssl_connection_manager.h
@@ -152,7 +152,7 @@ class SSLConnectionManager {
 				      size_t &written);
 
 	static IO::IO_RESULT sslWriteIOvec(Connection &target_ssl_connection,
-					   const iovec *__iovec, int count,
+					   const iovec *__iovec, size_t count,
 					   size_t &nwritten);
 	static IO::IO_RESULT handleWriteIOvec(Connection &target_ssl_connection,
 					      iovec *iov, size_t &iovec_size,
